+++ webui/module/Restore/src/Restore/Model/Restore.php
@@ -5,7 +5,7 @@
  * bareos-webui - Bareos Web-Frontend
  *
  * @link      https://github.com/bareos/bareos for the canonical source repository
- * @copyright Copyright (C) 2013-2019 Bareos GmbH & Co. KG (http://www.bareos.org/)
+ * @copyright Copyright (C) 2013-2022 Bareos GmbH & Co. KG (http://www.bareos.org/)
  * @license   GNU Affero General Public License (http://www.gnu.org/licenses/)
  *
  * This program is free software: you can redistribute it and/or modify
@@ -31,18 +31,11 @@ use Zend\InputFilter\InputFilterInterface;
 
 class Restore implements InputFilterAwareInterface
 {
-
-   protected $job;
-   protected $client;
-   protected $restoreclient;
-   protected $fileset;
-   protected $beforedate;
-
    protected $inputFilter;
 
    public function setInputFilter(InputFilterInterface $inputFilter)
    {
-      throw new \Exception("setInputFiler() not used");
+      throw new \Exception("setInputFilter() not used");
    }
 
    public function getInputFilter()
@@ -170,6 +163,24 @@ class Restore implements InputFilterAwareInterface
             'required' => true,
             'filters' => array(
                array('name' => 'StripTags'),
+               array('name' => 'StringTrim'),
+            ),
+            'validators' => array(
+               array(
+                  'name' => 'StringLength',
+                  'options' => array(
+                     'encoding' => 'UTF-8',
+                     'min' => 1,
+                     'max' => 128
+                  )
+               )
+            )
+         ));
+
+         $inputFilter->add(array(
+            'name' => 'pluginoptions',
+            'required' => false,
+            'filters' => array(
                array('name' => 'StringTrim'),
             ),
             'validators' => array(